structured-extractor/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "mcp-bank-structured-extractor"
version = "1.0.0"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115",
    "uvicorn[standard]",
    "orjson",
    "pydantic>=2.0",
    "httpx",
    "beautifulsoup4",
    "readability-lxml",
    "lxml",
    "jsonschema",
    "anthropic",
    "x402",
    "redis[asyncio]",
    "python-dotenv",
    "extruct",
    "w3lib",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "pytest-asyncio",
    "httpx",
    "ruff",
    "respx",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.ruff]
line-length = 100
target-version = "py312"
